Small refactor: Index data by `ResourceLocation`…
…instead of `String`. This allows for writing blockstates and such outside of
the `kdlycontent` namespace without making a separate `Writer`.

Also changes `pack.mcmeta`s to just be written files directly instead of going
through the whole `Writer` and `UniqueStrMap` process. (Because `UniqueStrMap`
was removed and `UniqueRLMap` doesn't work for this purpose)

Some things can be cleaned up as a result of this, some of which will be done
in directly upcoming commits, as well as some further down the line.